feat(webapp_v2): enhance documentation and component structure for better migration guidance
- Updated `CLAUDE.md` to include additional reading materials for building UI components and migrating CLJS pages to React.
- Introduced `CLJS_PATTERNS.md` for mapping ClojureScript patterns to React equivalents, aiding developers during migration.
- Added `COMPONENTS.md` to catalog reusable components, hooks, and services, promoting code reuse and consistency.
- Created `MIGRATION_CHECKLIST.md` to provide a step-by-step guide for migrating pages from CLJS to React.
- Refactored sidebar components and introduced new helper functions for better organization and maintainability.

These changes improve the overall documentation and structure of the codebase, facilitating smoother transitions for developers working on the migration from ClojureScript to React.
