--- webapp_v2/src/layout/Sidebar.module.css
@@ -1,156 +1,0 @@
-/* ── Nav link ──────────────────────────────────────────────── */
-
-.navLink {
-  border-radius: 6px;
-}
-
-/*
- * --mantine-color-gray-0 is overridden on the sidebar root element so that
- * Mantine's NavLink hover (which uses that variable) resolves to our subtle
- * white tint instead of the default near-white. No !important needed.
- */
-.navLink:hover {
-  background-color: rgba(255, 255, 255, 0.05);
-  color: rgba(255, 255, 255, 0.9);
-}
-
-.navLink:hover * {
-  color: inherit;
-}
-
-/* Active state — override Mantine's --nl-bg / --nl-color */
-.navLink[data-active],
-.navLink[aria-current='page'] {
-  background-color: rgba(255, 255, 255, 0.07);
-  color: #fff;
-  --nl-bg: rgba(255, 255, 255, 0.07);
-  --nl-color: #fff;
-}
-
-.navLink[data-active] *,
-.navLink[aria-current='page'] * {
-  color: inherit;
-}
-
-.navLink[data-active]:hover,
-.navLink[aria-current='page']:hover {
-  background-color: rgba(255, 255, 255, 0.1) !important;
-}
-
-.navLink:focus-visible {
-  outline: 2px solid rgba(255, 255, 255, 0.5);
-  outline-offset: 1px;
-}
-
-/* Blocked paid feature — dimmed, still clickable (redirects to upgrade) */
-.navLinkBlocked {
-  opacity: 0.45;
-}
-
-/* ── Collapsed icon button ─────────────────────────────────── */
-
-.iconBtn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 40px;
-  height: 40px;
-  border-radius: 6px;
-  color: rgba(255, 255, 255, 0.7);
-  cursor: pointer;
-  background: transparent;
-  border: none;
-  padding: 0;
-  transition: background-color 120ms ease, color 120ms ease;
-}
-
-.iconBtn:hover {
-  background-color: rgba(255, 255, 255, 0.05);
-  color: rgba(255, 255, 255, 0.9);
-}
-
-.iconBtn:focus-visible {
-  outline: 2px solid rgba(255, 255, 255, 0.5);
-  outline-offset: 1px;
-}
-
-.iconBtnActive {
-  background-color: rgba(255, 255, 255, 0.07);
-  color: #fff;
-}
-
-/* ── Profile disclosure button ─────────────────────────────── */
-
-.profileItem {
-  border-radius: 0;
-}
-
-/* ── Collapse button (full-width footer, right-aligned, border-top) ── */
-
-.collapseBtn {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: 6px;
-  width: 100%;
-  height: 40px;
-  padding: 0 16px;
-  border: none;
-  border-top: 1px solid rgba(255, 255, 255, 0.1);
-  background: transparent;
-  color: rgba(255, 255, 255, 0.6);
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: 600;
-  box-sizing: border-box;
-  flex-shrink: 0;
-  transition: background-color 120ms ease, color 120ms ease;
-}
-
-.collapseBtn:hover {
-  background-color: rgba(255, 255, 255, 0.05);
-  color: rgba(255, 255, 255, 0.9);
-}
-
-.collapseBtn:focus-visible {
-  outline: 2px solid rgba(255, 255, 255, 0.5);
-  outline-offset: -2px;
-}
-
-/* ── Skip link (visible on focus for keyboard users) ──────── */
-
-.skipLink {
-  position: fixed;
-  top: -9999px;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  padding: 6px 12px;
-  background: #3e63dd;
-  color: #fff;
-  font-size: 14px;
-  font-weight: 600;
-  border-radius: 6px;
-  z-index: 9999;
-  text-decoration: none;
-  white-space: nowrap;
-}
-
-.skipLink:focus {
-  top: 50%;
-  outline: none;
-  box-shadow: 0 0 0 3px rgba(62, 99, 221, 0.4);
-}
-
-/* ── Screen-reader-only text ──────────────────────────────── */
-
-.srOnly {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border-width: 0;
-}
